+++ src/main/kotlin/file/FileUtils.kt
@@ -3,6 +3,7 @@ package file
 import mu.KotlinLogging
 import java.nio.file.Files
 import java.nio.file.Path
+import java.nio.file.StandardCopyOption
 
 val log = KotlinLogging.logger {}
 
@@ -24,6 +25,33 @@ fun clearFile(it: Path) {
 		}
 	}
 }
+
+fun copyFolder(src: Path, dest: Path) {
+	try {
+		Files.walk(src)
+			.forEach { source ->
+				try {
+					val target = dest.resolve(src.relativize(source))
+					copyPath(source, target)
+				} catch (e: Exception) {
+					e.printStackTrace()
+				}
+			}
+	} catch (ex: Exception) {
+		ex.printStackTrace()
+	}
+}
+
+private fun copyPath(source: Path?, target: Path?) {
+	if (Files.isDirectory(source)) {
+		if (!Files.exists(target)) {
+			Files.createDirectory(target)
+		}
+	} else {
+		Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING)
+	}
+}
+
 
 private fun clearPathInternal(it: Path, dir: Path) {
 	if (it != dir) {
